SMRT-262 Editing script to be delete target groups unreleated to load balancers

+++ scripts/destroy_albs_created_via_kubernetes.sh
@@ -1,17 +1,16 @@
 #!/usr/bin/env bash
 
-./setup_aws_creds.sh $1 $2
+environment=$1
+vpc_id=$2
+./setup_aws_creds.sh $environment $vpc_id
 
 LoadBalancersInVpc=$(aws elbv2 describe-load-balancers | jq '[.LoadBalancers[] | select(.VpcId == "'"$vpc_id"'").LoadBalancerArn]' | jq -r '.[]')
 
 for LoadBalancerArn in $LoadBalancersInVpc
 do
-   echo $LoadBalancerArn
    MatchingTagOnALB=$(aws elbv2 describe-tags --resource-arns $LoadBalancerArn | grep 'scos.delete.on.teardown' | wc -l)
    if [ "${MatchingTagOnALB}" -gt 0 ]; then
-        ListenerArns=$(aws elbv2 describe-listeners --load-balancer-arn $LoadBalancerArn | jq '[.Listeners[].ListenerArn]' | jq -r '.[]')
-        TargetGroupArns=$(aws elbv2 describe-target-groups --load-balancer-arn $LoadBalancerArn | jq '[.TargetGroups[].TargetGroupArn]' | jq -r '.[]')
-        
+        ListenerArns=$(aws elbv2 describe-listeners --load-balancer-arn $LoadBalancerArn | jq '[.Listeners[].ListenerArn]' | jq -r '.[]')        
         for ListenerArn in $ListenerArns
         do
             echo "Deleteing.... $ListenerArn"
@@ -20,14 +19,17 @@ do
 
         echo "Deleting.... $LoadBalancerArn"
         aws elbv2 delete-load-balancer --load-balancer-arn $LoadBalancerArn
-
-        for TargetGroupArn in $TargetGroupArns
-        do
-            echo "Deleteing.... $TargetGroupArn"
-            aws elbv2 delete-target-group --target-group-arn $TargetGroupArn
-        done
-
    fi
+done
+
+TargetGroupsInVpc=$(aws elbv2 describe-target-groups | jq '[.TargetGroups[] | select(.VpcId == "'"$vpc_id"'").TargetGroupArn]' | jq -r '.[]')
+for TargetGroupArn in $TargetGroupsInVpc
+do
+    MatchingTagOnTargetGroup=$(aws elbv2 describe-tags --resource-arns $TargetGroupArn | grep 'scos.delete.on.teardown' | wc -l)
+    if [ "${MatchingTagOnTargetGroup}" -gt 0 ]; then
+        echo "Deleting.... $TargetGroupArn"
+        aws elbv2 delete-target-group --target-group-arn $TargetGroupArn
+    fi
 done
 
 exit 0
